Update README.md
This project aims to promote respect for copyright and encourage the responsible use of sheet music and music files. We recognize the importance of valuing and protecting the creative work of artists and musicians, and we want to raise awareness about the ethics of using the work of others.

In this repository, you will find relevant information on copyright, as well as recommendations for the responsible use of sheet music and music files. We also stress the importance of obtaining proper permissions when using copyrighted material and encourage users to search for content that is available for free use and does not infringe copyright.